cmd/snapd/daemon: move snapd daemon main into a subpackage
As a prerequisite step for merging snap and snapd binaries, move the
dameon code to a dedicated subpackage.
